requirements.txt
# State DOT Maintenance Manual RAG - Requirements

# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# PDF processing
PyMuPDF>=1.23.0

# Embeddings & ML
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector database
chromadb>=0.4.0

# LangChain ecosystem
langchain>=0.1.0
langchain-groq>=0.0.1
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langsmith>=0.1.0

# Utilities
tqdm>=4.66.0

